+++ src/main/java/net/more_rpg_classes/custom/CustomSpellImpacts.java
@@ -1,0 +1,20 @@
+package net.more_rpg_classes.custom;
+
+import net.minecraft.util.Identifier;
+import net.spell_engine.api.spell.event.SpellHandlers;
+
+import static net.more_rpg_classes.MRPGCMod.MOD_ID;
+
+public class CustomSpellImpacts {
+
+    public static void registerCustomImpacts(){
+        SpellHandlers.registerCustomImpact(
+                Identifier.of(MOD_ID, "knock_up"),
+                new KnockUpSpellImpact()
+        );
+        SpellHandlers.registerCustomImpact(
+                Identifier.of(MOD_ID, "stop_arrows"),
+                new StopArrowsImpact()
+        );
+    }
+}